requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
scipy>=1.7.0
tqdm>=4.60.0
# cupy-cuda11x>=10.0.0 # Optional: Uncomment and adjust cuda version (e.g., cupy-cuda12x) if using GPU acceleration
